changed Makefile (added ecdsa bundle) and deleted parts of ecdsa test
